start using a factory for the delete-modal stores
